fix(contracts): inline agent-profile so plain-node consumers can load the package
The contracts package is consumed as raw TS source by plain Node (type
stripping) on customer VPS gateways, where nodenext-style ./module.js
relative specifiers do not resolve to .ts files. The agent-profile
re-export crashed gateway startup on deploy of v2026.07.13-744 and the
sync agent rolled the fleet back. Inline the schema into index.ts
(reusing the existing SAFE_REFERENCE/textEncoder helpers), delete the
separate module, and document the no-relative-imports invariant at the
former re-export site.
